Install git and mercurial
To make "go get" work in most cases you need either git or mercurial installed to clone the dependencies.
Now the golang class will install these but the requirement to update aptitude (on debian based machines) is not a concern for the golang class and so is handled outside of the class.